+++ src/main/java/ch/ethz/idsc/amodeus/routing/DefaultAStarLMRouter.java
@@ -49,12 +49,10 @@ public class DefaultAStarLMRouter implements AVRouter {
         @Named(AVModule.AV_MODE)
         TravelTime travelTime;
 
-
-
         @Override
         public AVRouter createRouter(RouterConfig routerConfig, Network network) {
             return new DefaultAStarLMRouter(DefaultParallelLeastCostPathCalculator.//
-                    create((int) config.getNumberOfThreads(), new FastAStarLandmarksFactory(), network, //
+                    create(config.getNumberOfThreads(), new FastAStarLandmarksFactory(), network, //
                             new OnlyTimeDependentTravelDisutilityFixed(travelTime), travelTime));
 
         }
