+++ src/test/java/ch/ethz/idsc/amodeus/lp/LPMinFlowTest.java
@@ -47,7 +47,7 @@ public class LPMinFlowTest {
         AVConfigGroup avCg = new AVConfigGroup();
         Config config = ConfigUtils.loadConfig(configFile.getAbsolutePath(), avCg);
         GeneratorConfig genConfig = avCg.getOperatorConfigs().values().iterator().next().getGeneratorConfig();
-        int numRt = (int) genConfig.getNumberOfVehicles();
+        int numRt = genConfig.getNumberOfVehicles();
         int endTime = (int) config.qsim().getEndTime();
         Scenario scenario = ScenarioUtils.loadScenario(config);
         network = scenario.getNetwork();
